+++ test/test_helper.rb
@@ -29,9 +29,7 @@ class Minitest::Test
   end
 
   def stub_patch_request(record, updated_keys, table: @table1, status: 200, headers: {}, return_body: nil)
-    return_body ||= {
-      fields: record.fields
-    }
+    return_body ||= { fields: record.fields }
     return_body = return_body.to_json
 
     request_body = {